feat: VideoPlayerViewController Video 재생 기능 구현

+++ VideoRecorder/Extensions/Double+Extension.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 
-extension Int {
+extension Double {
     func convertToTimeFormat() -> String {
         let formatter = DateComponentsFormatter()
         formatter.allowedUnits = [.minute, .second]
